+++ K2D2Project/Controller/Docks/Pilots/TurnTo.cs
@@ -11,33 +11,41 @@ namespace K2D2.Controller.Docks.Pilots;
 /// <summary>
 /// rotation used for docking
 /// </summary>
-public class TurnTo : ExecuteController
+public class DockingTurnTo : ExecuteController
 {
-    Vector3 wanted_dir = Vector3.zero;
 
     KSPVessel current_vessel;
 
-    public double angle;
+    public float angle;
+    public float max_angle;
 
     public enum Mode
     {
         Off,
         RetroSpeed,
+        TargetDock
     }
 
     public Mode mode = Mode.Off;
 
-    public void StartRetroSpeed()
+    public void StartRetroSpeed(float max_angle = 3)
     {
         mode = Mode.RetroSpeed;
+        this.max_angle = max_angle;
         Start();
     }
 
-  
+    public void StartDockAlign(float max_angle = 1)
+    {
+        mode = Mode.TargetDock;
+        this.max_angle = max_angle;
+        Start();
+    }
+
     public override void Start()
     {
         current_vessel = K2D2_Plugin.Instance.current_vessel;
-         
+
         // reset time warp
         TimeWarpTools.SetRateIndex(0, false);
         var autopilot = current_vessel.Autopilot;
@@ -45,7 +53,7 @@ public class TurnTo : ExecuteController
         autopilot.SetMode(AutopilotMode.StabilityAssist);
     }
 
-    void CheckRetroSpeed()
+    void UpdateRetroSpeed()
     {
         var autopilot = current_vessel.Autopilot;
 
@@ -58,35 +66,89 @@ public class TurnTo : ExecuteController
 
         autopilot.SAS.SetTargetOrientation(direction, false);
 
+        finished = false;
+
         if (!checkRetroSpeed())
             return;
 
         if (!checkAngularRotation())
             return;
-        
+
+        finished = true;
+    }
+
+    Vector wanted_direction;
+
+    void UpdateTargetDock()
+    {
+        var autopilot = current_vessel.Autopilot;
+
+        // force autopilot
+        autopilot.Enabled = true;
+        autopilot.SAS.lockedMode = false;
+
+
+        var target = current_vessel.VesselComponent.TargetObject;
+        if (target == null)
+        {
+            mode = Mode.Off;
+        }
+
+        wanted_direction = current_vessel.VesselComponent.TargetObject.transform.up;
+        wanted_direction.vector = -wanted_direction.vector;
+
+        autopilot.SAS.SetTargetOrientation(wanted_direction, false);
+
+        finished = false;
+
+        if (!checkTargetDock())
+            return;
+
+        if (!checkAngularRotation())
+            return;
+
+        finished = true;
     }
 
     public override void Update()
     {
-        switch(mode) 
+        switch(mode)
         {
-            case Mode.RetroSpeed: CheckRetroSpeed(); break;
-        }   
+            case Mode.RetroSpeed: UpdateRetroSpeed(); break;
+            case Mode.TargetDock: UpdateTargetDock(); break;
+        }
     }
 
     public bool checkRetroSpeed()
     {
-        double max_angle = 5;
-
+        var control_component = current_vessel.VesselComponent.GetControlOwner();
         Vector retro_dir = current_vessel.VesselComponent.TargetVelocity;
-        Rotation vessel_rotation = current_vessel.GetRotation();
+        Rotation control_rotation = control_component.transform.Rotation;
 
         // convert rotation to speed coordinates system
-        vessel_rotation = Rotation.Reframed(vessel_rotation, retro_dir.coordinateSystem);
+        control_rotation = Rotation.Reframed(control_rotation, retro_dir.coordinateSystem);
 
-        Vector3d forward_direction = (vessel_rotation.localRotation * Vector3.up).normalized;
+        Vector3d forward_direction = (control_rotation.localRotation * Vector3.down).normalized;
 
         angle = (float)Vector3d.Angle(retro_dir.vector, forward_direction);
+        status_line = $"Waiting for good sas direction\nAngle = {angle:n2}°";
+
+        return angle < max_angle;
+    }
+
+    public bool checkTargetDock()
+    {
+        var control_component = current_vessel.VesselComponent.GetControlOwner();
+
+        // Vector retro_dir = current_vessel.VesselComponent.TargetVelocity;
+        Rotation control_rotation = control_component.transform.Rotation;
+
+        // convert rotation to speed coordinates system
+        control_rotation = Rotation.Reframed(control_rotation, wanted_direction.coordinateSystem);
+
+        Vector3d forward_direction = (control_rotation.localRotation * Vector3.down).normalized;
+
+        angle = (float)Vector3d.Angle(wanted_direction.vector, forward_direction);
         status_line = $"Waiting for good sas direction\nAngle = {angle:n2}°";
 
         return angle < max_angle;
@@ -128,10 +190,8 @@ public class TurnTo : ExecuteController
             // var angulor_vel_coord = VesselInfos.GetAngularSpeed().coordinateSystem;
             var angularVelocity = current_vessel.GetAngularSpeed().vector;
 
-            
             UI_Tools.Console($"angle {angle:n2} °");
             UI_Tools.Console($"angularVelocity {StrTool.Vector3ToString(angularVelocity)}");
-            
             UI_Tools.Console($"autopilot {autopilot.AutopilotMode}");
         }
     }